+++ packages/plugin-app-react-webpack/src/actions/build.ts
@@ -6,7 +6,7 @@ import { Action } from '@inventorjs/core'
 import webpack, { type Configuration } from 'webpack'
 import webpackFactory from '../config/webpackFactory.js'
 
-const CRITICAL_SIZE = 1024 * 200
+const CRITICAL_SIZE = 1024 * 244
 export default class extends Action {
   description = '构建项目'
   options = []
@@ -34,22 +34,30 @@ export default class extends Action {
           }) ?? {}
 
         if (statJson.errors?.length) {
-          this.log.raw(statJson.errors.map((item) => item.message).join('\n'))
+          this.log.error('Compile with errors:')
+          this.log.error(statJson.errors.map((item) => item.message).join('\n'))
           reject(statJson.errors)
           return
         }
         if (statJson.warnings?.length) {
-          this.log.error('Compile with warnings.')
-          this.log.raw(statJson.warnings.map((item) => item.message).join('\n'))
+          this.log.warn('Compile with warnings:')
+          this.log.raw(
+            this.color.yellow(
+              statJson.warnings
+                .filter((item) => !item.message.includes('size limit:'))
+                .map((item) => item.message)
+                .join('\n'),
+            ),
+          )
         }
         this.log.raw(
           statJson.assets?.map((asset) => {
             const humanSize = this.util.humanSize(asset.size)
-            const criticalSize = this.util.humanSize(CRITICAL_SIZE)
+            const humanCritical = this.util.humanSize(CRITICAL_SIZE)
             return asset.size > CRITICAL_SIZE
               ? [
-                  this.color.cyan(asset.name),
-                  this.color.red(`${humanSize}[exceed ${criticalSize}]`),
+                  this.color.red(asset.name),
+                  this.color.red(`${humanSize}[exceeds ${humanCritical}]`),
                 ]
               : [this.color.cyan(asset.name), this.color.cyan(humanSize)]
           }),