+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @inventorjs/cli
 
+## 3.0.5
+
+### Patch Changes
+
+- 77912ca: change dev port
+- Updated dependencies [77912ca]
+  - @inventorjs/core@3.11.5
+  - @inventorjs/plugin-app@3.0.5
+  - @inventorjs/plugin-plugin@3.0.5
+
 ## 3.0.4
 
 ### Patch Changes